+++ src/frontend/js/components/Root/index.tsx
@@ -9,6 +9,7 @@ import { Spinner } from 'components/Spinner';
 import ErrorBoundary from 'utils/errors/ErrorBoundary';
 import context from 'utils/context';
 
+const CourseProductItem = lazy(() => import('components/CourseProductItem'));
 const CourseRunEnrollment = lazy(() => import('components/CourseRunEnrollment'));
 const LanguageSelector = lazy(() => import('components/LanguageSelector'));
 const LtiConsumer = lazy(() => import('components/LtiConsumer'));
@@ -22,6 +23,7 @@ const Dashboard = lazy(() => import('components/Dashboard'));
 // for type-safety when we call them. This will let us use the props for any top-level component in a
 // way TypeScript understand and accepts
 interface ComponentLibrary {
+  CourseProductItem: typeof CourseProductItem;
   CourseRunEnrollment: typeof CourseRunEnrollment;
   Dashboard: typeof Dashboard;
   LanguageSelector: typeof LanguageSelector;
@@ -33,6 +35,7 @@ interface ComponentLibrary {
 }
 // Actually create the component map that we'll use below to access our component classes
 const componentLibrary: ComponentLibrary = {
+  CourseProductItem,
   CourseRunEnrollment,
   Dashboard,
   LanguageSelector,