+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
+- Refactor CourseProductItem as Joanie products are Richie course runs
 - Use versionned Joanie api routes
 - Migrate to React 18 and React-Query v4
 - Migrate to useResources hooks